4.15-4.23 complete, short some extra tests

+++ controllers/blogs.js
@@ -2,15 +2,16 @@ require('dotenv').config()
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
-const jwt = require('jsonwebtoken')
+//const jwt = require('jsonwebtoken')
+const userExtractor = require('../utils/userExtractor')
 
-const getTokenFrom = request => {
+/*const getTokenFrom = request => {
   const authorization = request.get('authorization')
   if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
     return authorization.substring(7)
   }
   return null
-}
+}*/
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({})
@@ -28,16 +29,23 @@ blogsRouter.get('/:id', async (request, response) => {
   }
 })
 
-blogsRouter.post('/', async (request, response) => {
+blogsRouter.post('/', userExtractor, async (request, response) => {
   const body = request.body
 
-  const token = getTokenFrom(request)
-  const decodedToken = jwt.verify(token, process.env.SECRET)
-  if (!decodedToken.id) {
+  //const token = getTokenFrom(request)
+  //const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  //if (!decodedToken.id) {
+  if (!request.user) {
     return response.status(401).json({ error: 'token missing or invalid' })
   }
-  const user = await User.findById(decodedToken.id)
+  //console.log(decodedToken)
+  const user = await User.findById(request.user)
+  if(!user) {
+    return response.status(401).json({ error: 'user does not exist' })
+  }
+
   //const user = await User.findById(body.userId)
+  //if(!user._id)
 
   const blog = new Blog({
     title: body.title,
@@ -53,21 +61,61 @@ blogsRouter.post('/', async (request, response) => {
   response.status(201).json(savedBlog)
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
+blogsRouter.delete('/:id', userExtractor, async (request, response) => {
+  //const token = getTokenFrom(request)
+  //const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  //if (!decodedToken.id) {
+  if (!request.user) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+  //const user = await User.findById(decodedToken.id)
+  const user = await User.findById(request.user)
+  if(!user) {
+    return response.status(401).json({ error: 'user does not exist' })
+  }
+
+  const blog = await Blog.findById(request.params.id)
+    .populate('user', { _id: 1 })
+  if(!blog){
+    return response.status(204).json({ error: 'blog not found' })
+  }
+
+  //console.log('blog to delete', blog)
+  if (request.user !== blog.user.id) {
+    return response.status(401).json({ error: 'user not authorized' })
+  }
+
   await Blog.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
 
-blogsRouter.put('/:id', async (request, response) => {
-  const body = request.body
+blogsRouter.put('/:id', userExtractor, async (request, response) => {
 
+  if (!request.user) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+
+  const user = await User.findById(request.user)
+  if(!user) {
+    return response.status(401).json({ error: 'user does not exist' })
+  }
+
+  const blogToUpdate = await Blog.findById(request.params.id)
+    .populate('user', { _id: 1 })
+  if(!blogToUpdate){
+    return response.status(204).json({ error: 'blog not found' })
+  }
+
+  if (request.user !== blogToUpdate.user.id) {
+    return response.status(401).json({ error: 'user not authorized' })
+  }
+  const body = request.body
   const blog = {
     title: body.title,
     author: body.author,
     url: body.url,
     likes: body.likes
   }
-
   const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, { new: true })
   response.json(updatedBlog)
   //console.log(response.json(updatedBlog))
